fuel_gauge: Move SBS API definions
The fuel gauge API header contains definitions that are exceptionally
specific to the SBS spec. While these SBS specific properties are useful to
have, they clutter up the main fuel gauge API and may semantically confuse
clients as to whether the API is SBS specific or not.

Move the SBS properies and definitons to a different header.
